major refactoring removing train/models:
- model classes deleted
- use model_vars directly in estimator
- model_vars contains model from models/ submodule and can access its attributes
- duplicate param_bounds together with ProcessModel class deleted as it is obsolote now
- imports cleaned and fixed
