+++ batchglm/train/numpy/base_glm/__init__.py
@@ -1,4 +1,2 @@
 from .estimator import EstimatorGlm
-from .model import ModelIwls
-from .processModel import ProcessModelGlm
 from .vars import ModelVarsGlm
