+++ batchglm/train/numpy/base_glm/estimator.py
@@ -12,9 +12,7 @@ import scipy.optimize
 import scipy.sparse
 import sparse
 
-from ....models.base_glm import InputDataGLM
 from .external import _EstimatorGLM, pkg_constants
-from .model import ModelIwls
 from .training_strategies import TrainingStrategies
 
 logger = logging.getLogger("batchglm")
@@ -33,22 +31,18 @@ class EstimatorGlm(_EstimatorGLM, metaclass=abc.ABCMeta):
 
     def __init__(
         self,
-        model: ModelIwls,
-        input_data: InputDataGLM,
         dtype: str,
     ):
         """
         Performs initialisation and creates a new estimator.
         :param model:
             The IWLS model to be fit
-        :param input_data:
-            The input data for the model
         :param dtype:
             i.e float64
         """
-        if input_data.design_scale.shape[1] != 1:
+        if self.model_vars.design_scale.shape[1] != 1:
             raise ValueError("cannot model more than one scale parameter with numpy backend right now.")
-        _EstimatorGLM.__init__(self=self, model=model, input_data=input_data)
+        _EstimatorGLM.__init__(self=self, model=model)
         self.dtype = dtype
         self.lls = []
 
@@ -103,9 +97,9 @@ class EstimatorGlm(_EstimatorGLM, metaclass=abc.ABCMeta):
         else:
             update_scale_freq = np.inf
         epochs_until_scale_update = update_scale_freq
-        fully_converged = np.tile(False, self.model.model_vars.n_features)
+        fully_converged = np.tile(False, self.model_vars.num_features)
 
-        ll_current = -self.model.ll_byfeature.compute()
+        ll_current = -self.model_vars.ll_byfeature.compute()
         ll_last_scale_update = ll_current.copy()
         # logging.getLogger("batchglm").info(
         sys.stdout.write("iter   %i: ll=%f\n" % (0, np.sum(ll_current)))
@@ -126,16 +120,16 @@ class EstimatorGlm(_EstimatorGLM, metaclass=abc.ABCMeta):
                         nproc=nproc,
                     )
                     # Perform trial update.
-                    self.model.theta_scale = self.model.theta_scale + b_step
+                    self.model_vars.theta_scale = self.model_vars.theta_scale + b_step
                     # Reverse update by feature if update leads to worse loss:
-                    ll_proposal = -self.model.ll_byfeature_j(j=idx_update).compute()
+                    ll_proposal = -self.model_vars.ll_byfeature_j(j=idx_update).compute()
                     idx_bad_step = idx_update[np.where(ll_proposal > ll_current[idx_update])[0]]
-                    if isinstance(self.model.theta_scale, dask.array.core.Array):
-                        theta_scale_new = self.model.theta_scale.compute()
+                    if isinstance(self.model_vars.theta_scale, dask.array.core.Array):
+                        theta_scale_new = self.model_vars.theta_scale.compute()
                     else:
-                        theta_scale_new = self.model.theta_scale.copy()
+                        theta_scale_new = self.model_vars.theta_scale.copy()
                     theta_scale_new[:, idx_bad_step] = theta_scale_new[:, idx_bad_step] - b_step[:, idx_bad_step]
-                    self.model.theta_scale = theta_scale_new
+                    self.model_vars.theta_scale = theta_scale_new
                 else:
                     ll_proposal = ll_current[idx_update]
                     idx_bad_step = np.array([], dtype=np.int32)
@@ -148,20 +142,20 @@ class EstimatorGlm(_EstimatorGLM, metaclass=abc.ABCMeta):
             else:
                 # IWLS step for location model:
                 # Compute update.
-                idx_update = self.model.idx_not_converged
+                idx_update = self.model_vars.idx_not_converged
                 if self._train_loc:
                     a_step = self.iwls_step(idx_update=idx_update)
                     # Perform trial update.
-                    self.model.theta_location = self.model.theta_location + a_step
+                    self.model_vars.theta_location += a_step
                     # Reverse update by feature if update leads to worse loss:
-                    ll_proposal = -self.model.ll_byfeature_j(j=idx_update).compute()
+                    ll_proposal = -self.model_vars.ll_byfeature_j(j=idx_update).compute()
                     idx_bad_step = idx_update[np.where(ll_proposal > ll_current[idx_update])[0]]
-                    if isinstance(self.model.theta_scale, dask.array.core.Array):
-                        theta_location_new = self.model.theta_location.compute()
+                    if isinstance(self.model_vars.theta_scale, dask.array.core.Array):
+                        theta_location_new = self.model_vars.theta_location.compute()
                     else:
-                        theta_location_new = self.model.theta_location.copy()
+                        theta_location_new = self.model_vars.theta_location.copy()
                     theta_location_new[:, idx_bad_step] = theta_location_new[:, idx_bad_step] - a_step[:, idx_bad_step]
-                    self.model.theta_location = theta_location_new
+                    self.model_vars.theta_location = theta_location_new
                 else:
                     ll_proposal = ll_current[idx_update]
                     idx_bad_step = np.array([], dtype=np.int32)
@@ -186,9 +180,9 @@ class EstimatorGlm(_EstimatorGLM, metaclass=abc.ABCMeta):
                     )
                     < pkg_constants.LLTOL_BY_FEATURE,
                 )
-                self.model.converged = np.logical_or(fully_converged, converged_f)
+                self.model_vars.converged = np.logical_or(fully_converged, converged_f)
                 ll_last_scale_update = ll_current.copy()
-                fully_converged = self.model.converged.copy()
+                fully_converged = self.model_vars.converged.copy()
             else:
                 # Update intermediate convergence in self.model.converged.
                 converged_f = np.logical_or(
@@ -199,8 +193,8 @@ class EstimatorGlm(_EstimatorGLM, metaclass=abc.ABCMeta):
                     )
                     < pkg_constants.LLTOL_BY_FEATURE,
                 )
-                self.model.converged = np.logical_or(self.model.converged, converged_f)
-                if np.all(self.model.converged):
+                self.model_vars.converged = np.logical_or(self.model_vars.converged, converged_f)
+                if np.all(self.model_vars.converged):
                     # All location models are converged. This means that the next update will be b model
                     # update and all remaining intermediate a model updates can be skipped:
                     epochs_until_scale_update = 0
@@ -214,7 +208,7 @@ class EstimatorGlm(_EstimatorGLM, metaclass=abc.ABCMeta):
                     (" " if train_step < 10 else "") + (" " if train_step < 100 else "") + str(train_step),
                     np.sum(ll_current),
                     np.mean(fully_converged) * 100,
-                    np.mean(self.model.converged) * 100,
+                    np.mean(self.model_vars.converged) * 100,
                     str(epochs_until_scale_update == update_scale_freq),
                     time.time() - t0,
                 )
@@ -233,19 +227,19 @@ class EstimatorGlm(_EstimatorGLM, metaclass=abc.ABCMeta):
         A single step in IWLS
         :return: np.ndarray  (inferred param x features)
         """
-        w = self.model.fim_weight_location_location_j(j=idx_update)  # (observations x features)
-        ybar = self.model.ybar_j(j=idx_update)  # (observations x features)
+        w = self.model_vars.fim_weight_location_location_j(j=idx_update)  # (observations x features)
+        ybar = self.model_vars.ybar_j(j=idx_update)  # (observations x features)
         # Translate to problem of form ax = b for each feature:
         # (in the following, X=design and Y=counts)
         # a=X^T*W*X: ([features] x inferred param)
         # x=theta: ([features] x inferred param)
         # b=X^T*W*Ybar: ([features] x inferred param)
-        xh = np.matmul(self.model.design_loc, self.model.constraints_loc)
+        xh = np.matmul(self.model_vars.design_loc, self.model_vars.constraints_loc)
         xhw = np.einsum("ob,of->fob", xh, w)
         a = np.einsum("fob,oc->fbc", xhw, xh)
         b = np.einsum("fob,of->fb", xhw, ybar)
 
-        delta_theta = np.zeros_like(self.model.theta_location)
+        delta_theta = np.zeros_like(self.model_vars.theta_location)
         if isinstance(delta_theta, dask.array.core.Array):
             # Need to copy to prevent returning an immutable view of the data
             delta_theta = delta_theta.compute().copy()
@@ -319,27 +313,27 @@ class EstimatorGlm(_EstimatorGLM, metaclass=abc.ABCMeta):
         :return:
         """
         iter = 0
-        theta_scale_old = self.model.theta_scale.compute()
-        converged = np.tile(True, self.model.model_vars.n_features)
+        theta_scale_old = self.model_vars.theta_scale.compute()
+        converged = np.tile(True, self.model_vars.num_features)
         converged[idx_update] = False
-        ll_current = -self.model.ll_byfeature.compute()
+        ll_current = -self.model_vars.ll_byfeature.compute()
         while np.any(np.logical_not(converged)) and iter < max_iter:
             idx_to_update = np.where(np.logical_not(converged))[0]
-            jac = np.zeros_like(self.model.theta_scale).compute()
+            jac = np.zeros_like(self.model_vars.theta_scale).compute()
             # Use mean jacobian so that learning rate is independent of number of samples.
             jac[:, idx_to_update] = (
-                self.model.jac_scale_j(j=idx_to_update).compute().T / self.model.input_data.num_observations
+                self.model_vars.jac_scale_j(j=idx_to_update).compute().T / self.model_vars.num_observations
             )
-            self.model.theta_scale_j_setter(
-                value=(self.model.theta_scale.compute() + lr * jac)[:, idx_to_update], j=idx_to_update
+            self.model_vars.theta_scale_j_setter(
+                value=(self.model_vars.theta_scale.compute() + lr * jac)[:, idx_to_update], j=idx_to_update
             )
             # Assess convergence:
             ll_previous = ll_current
-            ll_current = -self.model.ll_byfeature.compute()
+            ll_current = -self.model_vars.ll_byfeature.compute()
             converged_f = (ll_current - ll_previous) / ll_previous > -ftol
-            theta_scale_new = self.model.theta_scale.compute()
+            theta_scale_new = self.model_vars.theta_scale.compute()
             theta_scale_new[:, converged_f] = theta_scale_new[:, converged_f] - lr * jac[:, converged_f]
-            self.model.theta_scale = theta_scale_new
+            self.model_vars.theta_scale = theta_scale_new
             converged = np.logical_or(converged, converged_f)
             iter += 1
             logging.getLogger("batchglm").info(
@@ -350,7 +344,7 @@ class EstimatorGlm(_EstimatorGLM, metaclass=abc.ABCMeta):
                     np.mean(converged) * 100,
                 )
             )
-        return self.model.theta_scale.compute() - theta_scale_old
+        return self.model_vars.theta_scale.compute() - theta_scale_old
 
     def optim_handle(self, b_j, data_j, eta_loc_j, xh_scale, max_iter, ftol):
         # Need to supply dense numpy array to scipy optimize:
@@ -359,8 +353,8 @@ class EstimatorGlm(_EstimatorGLM, metaclass=abc.ABCMeta):
         if len(data_j.shape) == 1:
             data_j = np.expand_dims(data_j, axis=-1)
 
-        ll = self.model.ll_handle()
-        lb, ub = self.model.param_bounds(dtype=data_j.dtype)
+        ll = self.model_vars.ll_handle()
+        lb, ub = self.model_vars.param_bounds(dtype=data_j.dtype)
         lb_bracket = np.max([lb["theta_scale"], b_j - 20])
         ub_bracket = np.min([ub["theta_scale"], b_j + 20])
 
@@ -397,13 +391,13 @@ class EstimatorGlm(_EstimatorGLM, metaclass=abc.ABCMeta):
         A single loop step for the scale model
         :return:
         """
-        delta_theta = np.zeros_like(self.model.theta_scale)
+        delta_theta = np.zeros_like(self.model_vars.theta_scale)
         if isinstance(delta_theta, dask.array.core.Array):
             # Need to copy to prevent returning an immutable view of the data
             delta_theta = delta_theta.compute().copy()
 
-        xh_scale = np.matmul(self.model.design_scale, self.model.constraints_scale).compute()
-        theta_scale = self.model.theta_scale.compute()
+        xh_scale = np.matmul(self.model_vars.design_scale, self.model_vars.constraints_scale).compute()
+        theta_scale = self.model_vars.theta_scale.compute()
         if nproc > 1 and len(idx_update) > nproc:
             sys.stdout.write(
                 "\rFitting %i dispersion models: (progress not available with multiprocessing)" % len(idx_update)
@@ -411,7 +405,7 @@ class EstimatorGlm(_EstimatorGLM, metaclass=abc.ABCMeta):
             sys.stdout.flush()
             with multiprocessing.Pool(processes=nproc) as pool:
                 x = self.x.compute()
-                eta_loc = self.model.eta_loc.compute()
+                eta_loc = self.model_vars.eta_loc.compute()
                 results = pool.starmap(
                     self.optim_handle,
                     [(theta_scale[0, j], x[:, [j]], eta_loc[:, [j]], xh_scale, max_iter, ftol) for j in idx_update],
@@ -429,14 +423,14 @@ class EstimatorGlm(_EstimatorGLM, metaclass=abc.ABCMeta):
                 )
                 sys.stdout.flush()
                 if method.lower() == "brent":
-                    eta_loc = self.model.eta_loc_j(j=j).compute()
+                    eta_loc = self.model_vars.eta_loc_j(j=j).compute()
                     data = self.x[:, [j]].compute()
                     # Need to supply dense numpy array to scipy optimize:
                     if isinstance(data, sparse.COO) or isinstance(data, scipy.sparse.csr_matrix):
                         data = data.todense()
 
-                    ll = self.model.ll_handle()
-                    lb, ub = self.model.param_bounds(dtype=data.dtype)
+                    ll = self.model_vars.ll_handle()
+                    lb, ub = self.model_vars.param_bounds(dtype=data.dtype)
                     lb_bracket = np.max([lb["theta_scale"], theta_scale[0, j] - 20])
                     ub_bracket = np.min([ub["theta_scale"], theta_scale[0, j] + 20])
 
@@ -457,10 +451,10 @@ class EstimatorGlm(_EstimatorGLM, metaclass=abc.ABCMeta):
             sys.stdout.write("\r")
             sys.stdout.flush()
 
-        if isinstance(self.model.theta_scale, dask.array.core.Array):
-            delta_theta[:, idx_update] = delta_theta[:, idx_update] - self.model.theta_scale.compute()[:, idx_update]
+        if isinstance(self.model_vars.theta_scale, dask.array.core.Array):
+            delta_theta[:, idx_update] = delta_theta[:, idx_update] - self.model_vars.theta_scale.compute()[:, idx_update]
         else:
-            delta_theta[:, idx_update] = delta_theta[:, idx_update] - self.model.theta_scale.copy()[:, idx_update]
+            delta_theta[:, idx_update] = delta_theta[:, idx_update] - self.model_vars.theta_scale.copy()[:, idx_update]
         return delta_theta
 
     def finalize(self):
@@ -472,13 +466,13 @@ class EstimatorGlm(_EstimatorGLM, metaclass=abc.ABCMeta):
         transfers relevant attributes.
         """
         # Read from numpy-IRLS estimator specific model:
-        self._hessian = -self.model.fim.compute()
+        self._hessian = -self.model_vars.fim.compute()
         fisher_inv = np.zeros_like(self._hessian)
         invertible = np.where(np.linalg.cond(self._hessian, p=None) < 1 / sys.float_info.epsilon)[0]
         fisher_inv[invertible] = np.linalg.inv(-self._hessian[invertible])
         self._fisher_inv = fisher_inv
-        self._jacobian = np.sum(np.abs(self.model.jac.compute() / self.model.x.shape[0]), axis=1)
-        self._log_likelihood = self.model.ll_byfeature.compute()
+        self._jacobian = np.sum(np.abs(self.model_vars.jac.compute() / self.model_vars.x.shape[0]), axis=1)
+        self._log_likelihood = self.model_vars.ll_byfeature.compute()
         self._loss = np.sum(self._log_likelihood)
 
     @abc.abstractmethod
