+++ batchglm/train/numpy/glm_nb/__init__.py
@@ -1,4 +1,2 @@
 from .estimator import Estimator
-from .model import ModelIwlsNb
-from .processModel import ProcessModel
 from .vars import ModelVars
