+++ batchglm/train/numpy/glm_nb/external.py
@@ -1,10 +1,9 @@
 import batchglm.data as data_utils
 from batchglm import pkg_constants
-from batchglm.models.base_glm import InputDataGLM
 from batchglm.models.base_glm.utils import closedform_glm_mean, closedform_glm_scale
-from batchglm.models.glm_nb import Model, _EstimatorGLM
+
 from batchglm.models.glm_nb.utils import init_par
 
 # import necessary base_glm layers
-from batchglm.train.numpy.base_glm import EstimatorGlm, ModelIwls, ModelVarsGlm, ProcessModelGlm
+from batchglm.train.numpy.base_glm import EstimatorGlm, ModelVarsGlm
 from batchglm.utils.linalg import groupwise_solve_lm
